Move RequestLoggingMiddleware to standalone
